+++ R/find_conda.R
@@ -10,7 +10,7 @@
 #' @export
 #' @importFrom basilisk.utils getCondaBinary getCondaDir
 find_conda <- function(conda="auto",
-                       use_basilisk=FALSE){
+                       use_basilisk=TRUE){
     if(isTRUE(use_basilisk)){
         conda_x <- basilisk.utils::getCondaBinary(
             loc = basilisk.utils::getCondaDir()